Extract history classes and test into their own files
